pyproject.toml
[tool.poetry]
name = "starnavi-social-network"
version = "0.1.0"
description = ""
authors = ["Stanislav Buldyhin <faquedust@gmail.com>"]
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
python-json-logger = "^2.0.7"
orjson = "^3.9.10"
psycopg = {extras = ["binary"], version = "^3.1.14"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
